Adding protecting Video Touters and  UserDetail

+++ controllers/videoController.js
@@ -49,7 +49,6 @@ export const postUpload = async (req, res) => {
 };
 
 export const videoDetail = async (req, res) => {
-  // console.log(req.params);
   const {
     params: { id }
   } = req;
@@ -68,7 +67,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: video.title, video });
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routers.home);
@@ -94,7 +97,12 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findByIdAndDelete({ _id: id });
+    const video = await Video.findById(id);
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findByIdAndDelete({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
